feat: Refactor investment calculation logic into UserInputService and update UserInputComponent to use the service
